finish reorganizing jobsPage
