+++ client/src/components/jobsPage/jobListing/jobDetails/jobDetails.scss
@@ -1,4 +1,4 @@
-@import '../../App.scss';
+@import '../../../../App.scss';
 
 .details {
   font-size: large;
