+++ client/src/components/jobsPage/jobListing/jobListing.scss
@@ -1,4 +1,4 @@
-@import '../../App.scss';
+@import '../../../App.scss';
 
 .bp4-dialog {
   @media (max-width: $mobile) {
