+++ client/src/components/jobsPage/jobListing/jobListing.test.jsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react"
-import { UserProvider } from "../contexts/user";
+import { UserProvider } from "../../contexts/user";
 import JobListing from "./jobListing";
 import userEvent from '@testing-library/user-event';
