+++ client/src/components/jobsPage/sidebar/sidebar.scss
@@ -1,4 +1,4 @@
-@import '../../App.scss';
+@import '../../../App.scss';
 
 nav {
   display: flex;
